@@ -53,6 +53,7 @@ func NewCloudConfig(cloudName string) (config cloud.Configuration, err error) {
 	return
 }
 
+// injectServiceConfig injects a serviceconfiguration into cloud config
 func injectServiceConfig(config *cloud.Configuration, serviceName cloud.ServiceName, serviceConfig cloud.ServiceConfiguration) {
 	if config.Services == nil {
 		config.Services = map[cloud.ServiceName]cloud.ServiceConfiguration{}
@@ -61,6 +62,7 @@ func injectServiceConfig(config *cloud.Configuration, serviceName cloud.ServiceN
 	config.Services[serviceName] = serviceConfig
 }
 
+// createAzurePrivateCloudConfig creates azureprivate (onpremise) cloudconfig from either AZURE_CLOUD_CONFIG (string) or AZURE_CLOUD_CONFIG_FILE (file)
 func createAzurePrivateCloudConfig() (cloud.Configuration, error) {
 	var cloudConfigJson []byte
 	cloudConfig := cloud.Configuration{}
@@ -70,7 +72,7 @@ func createAzurePrivateCloudConfig() (cloud.Configuration, error) {
 		cloudConfigJson = []byte(val)
 	} else if val := os.Getenv("AZURE_CLOUD_CONFIG_FILE"); len(val) > 0 {
 		// cloud config via JSON file
-		data, err := os.ReadFile(val)
+		data, err := os.ReadFile(val) // #nosec G304
 		if err != nil {
 			return cloudConfig, fmt.Errorf(`unable to parse json for AzurePrivateCloud from env var AZURE_CLOUD_CONFIG_FILE, see https://github.com/webdevops/go-common/tree/main/azuresdk: %v`, err.Error())
 		}
